added license & throughput output

+++ src/core/perf.tests/aio.cpp
@@ -1,3 +1,36 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015 Microsoft Corporation
+ * 
+ * -=- Robust Distributed System Nucleus (rDSN) -=- 
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+/*
+ * Description:
+ *     Disk IO performance test
+ *
+ * Revision history:
+ *     2016-01-05, Tianyi Wang, first version
+ */
 #include <dsn/cpp/address.h>
 #include <dsn/internal/aio_provider.h>
 #include <gtest/gtest.h>
@@ -14,7 +47,7 @@ void aio_testcase(uint64_t block_size, size_t concurrency, bool is_write, bool r
     std::unique_ptr<char[]> buffer(new char[block_size]);
     std::atomic_int remain_concurrency;
     remain_concurrency = concurrency;
-    if (utils::filesystem::file_exists("temp"))
+    if (is_write && utils::filesystem::file_exists("temp"))
     {
         utils::filesystem::remove_path("temp");
         dassert(!utils::filesystem::file_exists("temp"), "");
@@ -72,7 +105,7 @@ TEST(core, aio_perf_test)
     auto block_size_list = { 64, 512, 4096, 512 * 1024 };
     auto concurrency_list = { 1, 4, 16 };
     for (auto is_write : { true, false }) {
-        for (auto random_offset : { false, true }) {
+        for (auto random_offset : {true, false }) {
             for (auto block_size : block_size_list)
             {
                 for (auto concurrency : concurrency_list)
